+++ packages/cashier/src/pages/account-transfer/account-transfer-form/index.js
@@ -1,0 +1,3 @@
+import AccountTransferForm from './account-transfer-form.jsx';
+
+export default AccountTransferForm;
