+++ packages/cashier/src/pages/account-transfer/account-transfer-locked/account-transfer-locked.jsx
@@ -6,7 +6,7 @@ import { Icon, Checklist, Text } from '@deriv/components';
 import { localize } from '@deriv/translations';
 import { connect } from 'Stores/connect';
 
-const TransferLocked = ({
+const AccountTransferLocked = ({
     is_financial_account,
     is_financial_information_incomplete,
     is_trading_experience_incomplete,
@@ -39,7 +39,7 @@ const TransferLocked = ({
     );
 };
 
-TransferLocked.propTypes = {
+AccountTransferLocked.propTypes = {
     is_financial_account: PropTypes.bool,
     is_financial_information_incomplete: PropTypes.bool,
     is_trading_experience_incomplete: PropTypes.bool,
@@ -49,4 +49,4 @@ export default connect(({ client }) => ({
     is_financial_account: client.is_financial_account,
     is_financial_information_incomplete: client.is_financial_information_incomplete,
     is_trading_experience_incomplete: client.is_trading_experience_incomplete,
-}))(TransferLocked);
+}))(AccountTransferLocked);
