+++ packages/cashier/src/pages/account-transfer/account-transfer-locked/index.js
@@ -1,0 +1,3 @@
+import AccountTransferLocked from './account-transfer-locked.jsx';
+
+export default AccountTransferLocked;
