+++ packages/cashier/src/pages/account-transfer/account-transfer-no-account/index.js
@@ -1,0 +1,3 @@
+import AccountTransferNoAccount from './account-transfer-no-account.jsx';
+
+export default AccountTransferNoAccount;
