+++ packages/cashier/src/pages/account-transfer/account-transfer-receipt/account-transfer-receipt.jsx
@@ -5,7 +5,7 @@ import { Button, Modal, Icon, Text } from '@deriv/components';
 import { formatMoney, getCurrencyDisplayCode, isMobile, routes } from '@deriv/shared';
 import { localize, Localize } from '@deriv/translations';
 import { connect } from 'Stores/connect';
-import 'Sass/account-transfer-receipt.scss';
+import './account-transfer-receipt.scss';
 
 const AccountTransferReceipt = ({
     disableApp,
