+++ packages/cashier/src/pages/account-transfer/account-transfer-receipt/account-transfer-receipt.scss
