+++ packages/cashier/src/pages/account-transfer/account-transfer-receipt/index.js
@@ -1,0 +1,3 @@
+import AccountTransferReceipt from './account-transfer-receipt.jsx';
+
+export default AccountTransferReceipt;
