+++ packages/cashier/src/pages/account-transfer/index.js
@@ -1,0 +1,3 @@
+import AccountTransfer from './account-transfer.jsx';
+
+export default AccountTransfer;
