+++ packages/cashier/src/pages/index.js
@@ -1,0 +1,3 @@
+import AccountTransfer from './account-transfer';
+
+export { AccountTransfer };
